Reduce startup overhead of pyxl being installed
pyxl needs to register itself as a codec on python startup, but it
doesn't need to actually import any of the pyxl machinery until it
needs to decode something.

The mild gymnastics performed (changing the module to import on
startup to `fast_register` and then importing that from the original
`register`) are to keep the diff size small while not breaking
`register`'s functionality as a script.

This shaves about 60ms off `python3 -c ''` with pyxl installed on my
laptop (from 100ms to 40ms!), which is 10ms more than a slightly
simpler change that just moves the `pyxl.codec.tokenizer` import into
`pyxl_transform`.