Improve the generated code for <if> conditions
Currently pyxl uses a global stack (!) of results from if conditions
to determine whether to execute else branches.

Instead, use if-else to do it in a more natural manner.

Unfortunately, because the parser is structured in a very streaming
manner, we don't know whether an if has an else block until after
fully processing it, so we do some tasteless rewriting of the output
to handle this.

The main purpose of this is to produce better looking code so that we
can compile everything to the target python and throw away the pyxl
codec, but I suppose it also fixes a thread-safety bug and probably
makes things faster too.