Emit 'str' instead of 'unicode' when joining together attr values
To the extent we care about this code actually working in Python 3,
this is obviously an improvement. We don't /really/ care about that,
so this is really more about making the typechecking work in both 2
and 3. (Using `str` on 2 is incorrect from an actual running code
perspective but should basically be fine from a mypy typechecking
perspective. We could use `typing.Text` or `six.string_types` but that
adds a dependency and would require an import in files that use it.)